Finished the First Section
Added our image and messed around with the CSS. Specifically adding .p-first since the styling from .header-link was messing with it. Then for the content we split it up into two sections (left & right) making sure to set the margin and width of the left side to match the example page.